More safely remove empty CHANGELOG entries (#1827)
Previous changelog empty entry removal was relying on an unsafe bash
side effect in order to work correctly. This in turn caused the
"correct" version of the script to fail when checked by "shell check".
This new version avoids the problem by narrowing the scope of sections
we are checking down to just those that are in the "unreleased" section
and then verifying that they are empty. Previously we checked "has
entry" and that was also more error prone than "has no entries". Seems
like it wouldn't be, but it is.

This has been tested extensively on OSX, Theo is testing Linux.
There are still issues herein where if the format deviates, it could go
bad, I opened an issue for those:

https://github.com/ponylang/ponyc/issues/1826

As they involve a much deeper rewrite/new approach.