Add support for custom serialisation (RFC 21)
This commit adds support for custom serialisation. Custom
serialisation lets the user provide methods that are called by Pony's
serialisation and deserialisation system in order to add additional
information that otherwise would not be serialized. The primary goal
of this system is to provide a means of serialising data stored in
`Pointer` member variables.