Properly account for foreign objects in GC
Prior to this commit, foreign objects were accounted for upon
receive but then their impact on the heap was lost whenever GC
ran. This commit resolves that by making sure that foreign objects
are properly accounted in the heap during the marking phase of GC
in a manner similar to how they're accounted for upon receive.

Resolves #1550