Fix JIT symbol load
When loading symbols from a JITed program in `gen_jit_and_run`, the
value stored is now the contents instead of the address. This avoids an
incorrect level of indirection.

In order to avoid copying big arrays, the compiler tests now load a
pointer to the generated descriptor table instead of the table itself.
The `__DescTable*` symbols are now `__PonyDescTable*` symbols to avoid
potential name conflicts.

Closes #1843.