Add iftype conditions (RFC 26) (#1855)
These new conditions allow programmers to get information on subtyping
relationships of type parameters at compile time to conditionally
compile code.

Most of the work is done during the scope pass, where the definition
assigned to a type parameter reference differs based on whether the
reference occurs in the then clause of an iftype condition.

This change implements only a part of RFC 26. Method specialisation is
still to be implemented.