Reify function references used as addressof operands correctly
Previously, a function reference used as the operand of `addressof` was
only reified in `expr_qualify`. Since this function is only called when
there is at least one type argument, a function referenced with no type
arguments wouldn't be checked. This change fixes the issue by reifying
the function in `expr_addressof` if the function reference wasn't
already reified.