Disallow exhaustive match when all cases jump away.
Theoretically, it would be possible to detect and take advantage of
exhaustive match when all exhaustive cases jump away, and treat the
entire match block as jumping away. Unfortunately there is a circular
dependency between the code that determines which expressions jump away
and the code that detects an exhaustive match.

For now, we simply disable exhaustive match detection in such cases.

In practice, you should still be able to do what you're trying to do -
if you move the "jump away" code to outside the match block, then
exhaustive match can be detected on the match block.

Resolves #1898.