Don't call TCPConnection backpressure notifies incorrectly (#1904)
It's possible under rare circumstances to call `throttled` and
`unthrottled` on a `TCPConnectionNotify` at the incorrect time. This
happens because we were not explicitly tracking whether our connection
was throttled or not.

If you were to be running without one shot enabled (unlikely now), this
code would spuriously call `unthrottled`. With this change, we no longer
have to worry that some other change might inadvertently start calling
`throttled` or `unthrottled` incorrectly.