Fix a bug in finaliser handling (#1908)
Previously, the finaliser bitmap was being checked against the slot
bitmap, and if the slot had not been freed, the loop continued. This
resulted in an infinite loop for a small chunk with allocations that
were both still in use and had finalisers. This change ANDs the
finaliser bitmap with the slot bitmap before running finalisers,
fixing this problem.