Change machine word constructors to have no default argument.
This is meant to prevent confusion with respect to other primitives
whose value can be compared using `is ${PrimitiveName}`. Machine words
are special primitives that can have multiple values, and comparing
with `is ${MachineWord}` is currently equivalent to comparing to zero.

Incidentally, fixing this also alerts us to the bug in #1920.

We should hold off on merging this PR until after #1927 is merged.