Use a constant-time algorithm for trait subtyping tests (#1752)
The previous algorithm for subtyping tests (used e.g. in pattern
matching) used an array of provided types in each type descriptor. The
array was iterated looking for the tested type, resulting in O(n)
complexity.

The algorithm implemented in this commit uses a bitmap of provided
types. A type provides a trait if the bit for that trait (determined
using the trait ID) is set. This results in O(1) complexity, with no
branching and a shorter code sequence.

The space complexity is now O(n) on the total number of traits instead
of the number of provided traits. The additional binary size is
mitigated by marking the traits data structures `unnamed_addr` during
code generation. This allows identical constants to be merged, avoiding
duplicates. Since a lot of types provide the same traits, a lot of
space is saved.