Prevent duplicate long_tests from being registered
Currently it is possible to call `long_test` more than once for a given
test. This results in more than 1 timer being registered, however, we
only keep track of 1 of timers registered so when the test is completed,
all timers are not cancelled and tests will continue "to run" until the
timer finishes.

This PR prevents duplicate long tests from being registered by only
setting up a long test if one hasn't already been registered. If one is
already registered, we skip the normal setup procedure and log a
warning.