Remove and replace misleading Time.wall_to_nanos
Based on its description and name, one would be reasonable in expecting
`Time.wall_to_nanos` to take a wall clock time and convert it the number
of nanoseconds since the "start of time". That isn't however what it
does. It was getting an absolute time to fire a timer setup in
Timer.abs.

This PR correct this problem by moving the functionality that previously
existed in `wall_to_nanos` into Timer as a private method and adds a
function to `Nanos` to convert from a wall clock time to nanoseconds.