Fix segfault in Array.trim_in_place (#1999)
This is the mirror update to Array.trim_in_place that was made
previously in String. Relevant commit comment from String:

Array.trim_in_place wasn't written to work correctly if you were
to trim the array down to nothing and then append a size less than the
"alloc" value that is left. This combination could result in incorrect
handling for arrays that end up with a size of 0.

This is because, a new pointer was created, but we retained incorrect
alloc info. This wouldn't cause an error so long as when you used the
array for something like an append, reserve allocated new memory.
However, if the appended value was less than the perceived alloc size,
no additional memory would be added and kaboom.

see 381b2816c6581dd2441f879d81ed9890ca181f4e