Remove the old lambda syntax from the compiler
In version 0.9, the concise lambda syntax from RFC 24 was implemented,
removing the `lambda ... end` syntax from the language. This syntax was
kept in the compiler to help report a helpful compiler error to aid
users in the transition.

Version 0.9 happened a while ago and users had time to adapt. Removing
the syntax altogether (treating old lambdas as plain parsing errors)
cleans up the compiler and doesn't have ill effects at that point.