Test for grammar changes in CI (#2034)
Several times in the past, the grammar rules in the parser were
modified without the pony.g file being updated. This change aims to
safeguard against this by testing whether the output of `ponyc --antlr`
is equal to the current pony.g file.

A mention of grammar changes and the pony.g file was also added to
CONTRIBUTING.md.