Forbid returning and passing tuples to FFI functions (#2012)
Passing complex objects by value in C is hard: there are many different
calling conventions depending on the platform. For example, a C
compiler is allowed to split the members of an object into different
parameters, or to change the return value into an out parameter.

The Pony compiler currently uses a naive implementation to pass objects
to FFI functions and isn't able to use the correct calling convention
for every call. Ideally, we'd want to rely on a C compiler API (like
LibClang) to generate the correct signatures and calls.

This change is a temporary measure to avoid spurious FFI bugs until
we have the interoperability detailed above. LLVM intrinsics are still
allowed to return and be passed tuples since their signatures are
defined by LLVM.