pony_become() the null actor on program termination
This change stops an instance of a Pony program from retaining a
reference to one of the actors it created. This prevents compiler JIT
tests from having a dangling reference to an actor from the previous
JIT test.

This also reverts 6f0c33a and a part of 1e1c484 since these changes
fixed symptoms of the problem and not the problem itself.