Break down the local block at pool allocator cleanup
This change breaks down the thread-local free block into thread-local
free list elements when cleaning up the memory pool-allocated by a
thread instead of pushing the whole block in a specific global list.

This allows the removal of an allocation case in pool_get and reduces
the amount of global free block lists to one.