Add copy_cpointer to String class to copy fixed number of bytes. (#2044)
The exisiting copy_cstring method could be used for either
null-terminated or fixed size strings. The string was considered
null-terminated when the given length was 0, leading to surprising
behaviour when copying a 0 sized string.

Instead the method is split in two, copy_cstring for null-terminated
and copy_cpointer for fixed size, following the existing naming
convention used by other methods.