Add ability to default enable PIC when compiling ponyc (#2113)
This commit allows folks compiling from source (for packaging or
for their own use) to easily specify that `ponyc` should by default
compile programs using Position Independent Code (PIC). The main
driving force behind this is a smoother user experience for folks
on distributions where PIC is enabled by default so they don't
have to always remember to pass `--pic` to `ponyc` when compiling
programs. Alpine Edge is one such example. With this change,
packaging for Alpine Edge is as simple as `make default_pic=true`
and then users of `ponyc` on Alpine Edge don't need to be aware
of PIC being enabled by default at all and don't have to remember
to pass `--pic` to every invocation of `ponyc`.

This commit also adds the ability to specify `--nopic` to disable
PIC if `ponyc` was compiled with `default_pic=true`.

NOTE: This does not change the default behavior of `ponyc` when
not compiling with `default_pic=true` and it is mainly a convenience
meant to be used only when the person compiling knows that the
toolset being used (gcc on a distro with PIC enabled by default)
will cause issues otherwise.