Fix cursor location for some reported errors and info. (#2136)
In PR #2039 (explicit partial calls), one of the changes I made was
to make it so that the cursor, when reporting on an error for a token
that was missing (like the missing '?' for an explicit partial call,
would be placed under the token to the left of where the missing
token should be, instead of under the token to the right (which
was often on the next line).

However, in doing this, I accidentally made it so that many other
(non-missing) tokens were also being reported one token to the left
instead of under the actual token itself. For example:

    main.pony:5:5: can't find declaration of 'present'
        until present end
        ^

Which should be displayed as:

    main.pony:5:5: can't find declaration of 'present'
        until present end
              ^

This PR fixes the accidental breakage, while retaining the desired
behaviour introduced in PR #2039. Now all tokens should be reported
in the correct position when displaying errors.