Codegen: Use distinct memory representation types
This change adds a new distinction in object types in the code
generator between a type's memory representation type and value type.
For example, Bool previously used LLVM's `i8` and now uses LLVM's `i1`
as its value type and either `i8` or `i32` as its memory representation
type depending on the platform. This currently is the only type whose
value and memory types differ.

The motivation for this change is to improve Pony's compatiblity with
the C ABI, which has the same kind of distinctions (for example, Clang
uses the same LLVM types as above to express C's boolean type).