Change the signature of runtime functions to take a pony_ctx_t (#2172)
This changes the signature of the public runtime functions
pony_continuation and pony_triggergc to take the execution context
instead of an actor as a parameter. Since these functions are only safe
to use on the current actor, this signature change better conveys the
function contract and can avoid misuse.

This also adds assertion checks on public runtime functions expecting a
current actor.