Make actor continuations a build time option
The Encore project uses the Pony runtime and has implemented
continuations for actors. There's a small amount of code in the Pony
runtime to support these continuations. It's a fairly unobtrusive bit of
code except for one small point. There's a check for continuation in a
hot path when actors are handling messages. This commit makes actor
continuation support a compile time option.

To get previous actor continuation support, USE_ACTOR_CONTINUATIONS
should be defined. Support for this exists in the Makefile as a "use"
directive.