Don't print capabilities for type params when generating docs (#2184)
Per
https://github.com/ponylang/ponyc/issues/2145#issuecomment-321131037,
this shouldn't be allowed as its not legal Pony. As #2145 generally
comments on we only want legal code output in the documentation.

This addresses the specific issue in #2145, however, there are other
instances we might want to expand this to. I decided to limit to this
specifically as I think any change we make in other sections of docgen
would require a thorough auditing of output documentation to make sure
there weren't any unintended consquences.

Prior to this change, generated documentation for:

```pony
type Set[A: (Hashable #read & Equatable[A] #read)] is HashSet[A,
HashEq[A]]
```

would be

```pony
type Set[A: (Hashable #read & Equatable[A #read] #read)] is
  HashSet[A #read, HashEq[A #read] val] ref
```

now it is

```
type Set[A: (Hashable #read & Equatable[A] #read)] is
  HashSet[A, HashEq[A] val] ref
```