Disallow calling methods of a type union whose parameter names differ when using named arguments.
This fixes the bug noted in #394, where named parameters can be used with methods in a union type whose parameter names differ.

The change makes it a compile-time error to call a method on a union type using named arguments when the methods in the union types have differently-named parameters.