Fix race condition in kqueue event system implementation (#2193)
This fixes a race condition in the kqueue back-end. Previously, it was possible for an unsubscribed (i.e. ASIO_DISPOSABLE) to be sent as an event in between being disposed and being sent as an ASIO_DISPOSABLE event. This could cause an event that had been unsubscribed to receive a read or a write event immediately before a dispose event.