Add coalescing of message sends (#2200)
This change adds the ability to send multiple messages in a single
message queue operation. This includes two modifications to the public
runtime interface:

- A new `pony_chain` function that chains two messages together. A
  chain with more than two elements can be built with successive calls
  to `pony_chain`. This function doesn't send messages.
- A change to `pony_sendv` and `pony_sendv_single`. These functions now
  take the two ends of a message chain instead of a single message. It
  is still possible to send a single message by passing said message as
  both the start and the end of the chain.

Adding varargs to `pony_sendv` and `pony_sendv_single` and building the
chain internally has been considered as an alternative to the two step
process explained above, but was abandoned for performance reasons as
as varargs are very hard to optimise.

This also extends the MergeMessage optimisation pass to build message
chains.