Fix broken "make" command
My PR to introduce support for DragonFlyBSD added /usr/local/lib to the
linker search path. We assumed this would be harmless, however it was
not. Because /usr/local/lib was added to the path in a fashion where it
would appear before the local "build" directory then the following would
happen:

- Run `make`
- Get a linker failure if you'd previously installed ponyc

Whereas

- Run `make` install
- No linker error

The problem occured because the pony libraries in /usr/local/lib were
being used instead of those in build.

This PR moves the addition of /usr/local/lib to being AFTER build path.

Closes #2219