Fix small chunk finaliser bug (#2257)
Prior to this commit there was a bug which prevented the finaliser
from running for the first slot of a small `chunk` in an actors
`heap`. This would result in incorrect behavior but only on a subset
of all objects created depending on which slot in a small chunk the
object happened to occupy.

This commit fixes this issue and also adds a test to catch the
issue to prevent regression.