Rename Writer `reserve` to `reserve_current` and fix memory leak
Prior to this commit, multiple calls to `Writer.reserve` would
allocate more and more memory in `_current` even if a sufficient
amount of unused memory was already available and allocated in
`_current`.

This commit changes the logic in `reserve` to call `_check` as
that has similar logic to confirm/reserve memory required but
without the memory leak.

This commit also renames `reserve` to `reserve_current`.