docgen: don't break union types in headers
Before, many `from()` methods' docs were broken, because markdown
doesn't allow you to have line breaks in your headers, e.g.

    ### from[A: ((I8 val | I16 val | I32 val |
        [I64](builtin-I64) val | [I128](builtin-I128) val | [ILong](builtin-ILong) val |

Now, it'll not insert breaks for unions (introduced in #1294) when it is
a method header, but leaves it as one long line. (It certainly isn't
beautiful, but it is less broken...)

Turns out this bug has another instance in document headers:

    # Flag\[A: (([U8](builtin-U8) val | [U16](builtin-U16) val | [U32](builtin-U32) val |
        [U64](builtin-U64) val | [U128](builtin-U128) val | [ULong](builtin-ULong) val |

...will also be one long line now.

Also renames an existing `line_breaks` argument to `break_lines` to be
in-line with `generate_links`.

Signed-off-by: Stephan Renatus <srenatus@chef.io>