Implement RFC 47 (Serialise signature)
This change implements deterministic package and program signatures
based on the contents and dependencies of the packages. Signatures are
computed by hashing the serialised package ASTs with the BLAKE2 hash
function. This hash function was chosen for its good speed and low
collision rate.

The signature computation handles circular dependencies by computing
signatures per package cycle instead of per package. Packages in the
same group are discriminated by a group index.

In order to allow deterministic signatures, file parsing order is now
always alphabetical inside of a package.

The compiler now outputs the package groups and dependencies when the
`-V3` or `-V4` flags are enabled.

Signatures will be used for future work on separate package compilation
in order to validate package identities.

Closes #2147.