Garbage collect actors when --ponynoblock is in use (#2307)
Prior to this commit, actors wouldn't be garbage collected when using
--ponynoblock. With this change, any actors that aren't part of a cycle
will be garbage collected if they can no longer do work.

Before this change, using `--ponynoblock` with an application that
created transient actors would result in never ending memory growth as
the actors would never be gc'd. You can see an example of this by taking
a Pony program that listens for TCP connections and do a repeating cycle
of open/close. None of the created `TCPConnection` actors would be
garbage collected.

Closes #2306