Run next() for uninitialized Random values
The current implementations of XorOshiro128Plus and XorShift128Plus
return the sum of input seed values for the first value, which is
unsafe. Rather than telling people to discard the first value,
skip past it in the actual implementation.

This commit also fixes the documentation of current implementations
for consistency with the Random trait.

Resolves #2320