Fix alloc bug in String/Array trimming functions
Prior to this commit, the trim/trim_in_place functions of Array
and String incorrectly calculated the new `_alloc` value.

This commit changes the logic to set `_alloc` to be the same
as the new `_size` if the last byte is not included and to
only included extra allocated bytes if we're including the
last byte.

This commit also changes how `String.from_cpointer` calculates
the `_alloc` value as it was incorrectly adding a byte when it
shouldn't have been (this was likely due to an assumption of a
null terminator being present that doesn't hold any longer).

Tests have also been enhanced to prevent regressions.