Add `chop` function for chopping `iso` Strings and Arrays
This commit adds a `chop` function to both String and Array
that works on an `iso` and chops it in two with the left half
being a new `iso` String/Array reusing the same pointer
and the right half being a `trim_in_place` version of the original
String/Array. Both are returned as `iso` and mutable and do not
share any memory. The operation does not allocate a new pointer
nor copy elements allowing for efficiency.