Fix incorrect detection of exhaustive matches
for primitives that have custom 'eq' methods or implement Equatable[T] where T is a subtype of the primitive.

The type was inferred by looking at the type of the full case AST which was inferred
from the first argument of the eq method. Now the exhaustive match logic is
considering the actual type of the case expression.